Consider a dead oneshot service good
For a oneshot service, particular those started by a timer,
it is perfectly normal for a status
to be `dead`. e.g.

```
systemctl status mlocate-updatedb.service
● mlocate-updatedb.service - Update a database for mlocate
   Loaded: loaded (/usr/lib/systemd/system/mlocate-updatedb.service; static; vendor preset: disabled)
   Active: inactive (dead) since Mon 2020-06-15 00:00:10 CEST; 3 days ago
 Main PID: 28558 (code=exited, status=0/SUCCESS)
```

This patch checks the service type and considers a dead oneshot service
as a good service.

For completness when a oneshot command fails for whatever reason it appears as

```
● mlocate-updatedb.service - Update a database for mlocate
   Loaded: loaded (/usr/lib/systemd/system/mlocate-updatedb.service; static; vendor preset: disabled)
   Active: failed (Result: exit-code) since Thu 2020-06-18 18:58:29 CEST; 6s ago
  Process: 70720 ExecStart=/usr/libexec/mlocate-run-updated (code=exited, status=203/EXEC)
 Main PID: 70720 (code=exited, status=203/EXEC)
```

and this `failed` reports as an error status to collectd.