Consider state start as a good state
A service that is being started is a good state and should
be considered so by this sensor.

A service will stay in start state until TimeoutStartSec= is
reached. Default for this timeout is 90 seconds but infinity
for a oneshot service which will remain in start
state forever by default.